refactor: rename _id to user_id in Users model and repository for consistency
